ch(validation-fixes). - move validation code to new file [Fixes #161873577]
